# apps/api/swagger.yml
openapi: 3.0.0
paths:
...
  /v1/cars/{id}:
    get:
      operationId: CarsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car (UUID)
          schema:
            type: string
      responses:
        "200":
          description: Car successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_NOT_FOUND
                    message: Car not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Get a car
      tags:
        - Cars
    put:
      operationId: CarsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car (UUID)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarDto"
      responses:
        "200":
          description: Car was successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                CAR_NOT_FOUND:
                  summary: CAR_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_NOT_FOUND
                    message: Car not found
                CAR_MODEL_NOT_FOUND:
                  summary: CAR_MODEL_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MODEL_NOT_FOUND
                    message: Car model not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Update a car
      tags:
        - Cars
      security:
        - apiKey: []
    delete:
      operationId: CarsController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car (UUID)
          schema:
            type: string
      responses:
        "204":
          description: Car deleted successfully
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_NOT_FOUND
                    message: Car not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Delete a car
      tags:
        - Cars
      security:
        - apiKey: []
...
  /v1/car-models:
    post:
      operationId: CarModelsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarModelDto"
      responses:
        "201":
          description: Car model created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarModelResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car manufacturer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MANUFACTURER_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MANUFACTURER_NOT_FOUND
                    message: Car manufacturer not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Create a car model
      tags:
        - CarModels
      security:
        - apiKey: []
    get:
      operationId: CarModelsController_findAll_v1
      parameters:
        - name: sortDirection
          required: false
          in: query
          description: Sort direction (ascending or descending)
          schema:
            type: string
        - name: sortField
          required: false
          in: query
          description: Field to sort car models by
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: The limit of items to be returned.
          schema:
            type: number
            default: 20
        - name: skip
          required: false
          in: query
          description: The amount of items to skip.
          schema:
            type: number
            default: 0
        - name: manufacturerSlug
          required: false
          in: query
          description: The slug of the car manufacturer
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        "200":
          description: List of car models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarModelResponseListDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: List car models
      tags:
        - CarModels
  /v1/car-models/{id}:
    get:
      operationId: CarModelsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car model (UUID)
          schema:
            type: string
      responses:
        "200":
          description: Car model successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarModelResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "404":
          description: Car model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MODEL_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MODEL_NOT_FOUND
                    message: Car model not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Get a car model
      tags:
        - CarModels
    put:
      operationId: CarModelsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car model (UUID)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarModelDto"
      responses:
        "200":
          description: Car model was successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarModelResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MODEL_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MODEL_NOT_FOUND
                    message: Car model not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Update a car model
      tags:
        - CarModels
      security:
        - apiKey: []
    delete:
      operationId: CarModelsController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car model (UUID)
          schema:
            type: string
      responses:
        "204":
          description: Car model deleted successfully
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MODEL_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MODEL_NOT_FOUND
                    message: Car model not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Delete a car model
      tags:
        - CarModels
      security:
        - apiKey: []
  /v1/car-models/slug/{slug}:
    get:
      operationId: CarModelsController_findBySlug_v1
      parameters:
        - name: slug
          required: true
          in: path
          description: The slug of the car model
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        "200":
          description: Car model succesfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarModelResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "404":
          description: Car model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MODEL_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MODEL_NOT_FOUND
                    message: Car model not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Get a car model by it's slug
      tags:
        - CarModels
info:
  title: 🔥 Next Gen Nestjs API
  description: Demo API to showcase autogenerated Swagger docs from Zod schemas and tRPC.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ErrorDto:
      type: object
      properties:
        statusCode:
          description: The HTTP status code
          type: number
          enum:
            - 100
            - 101
            - 102
            - 103
            - 200
            - 201
            - 202
            - 203
            - 204
            - 205
            - 206
            - 207
            - 208
            - 210
            - 300
            - 301
            - 302
            - 303
            - 304
            - 307
            - 308
            - 400
            - 401
            - 402
            - 403
            - 404
            - 405
            - 406
            - 407
            - 408
            - 409
            - 410
            - 411
            - 412
            - 413
            - 414
            - 415
            - 416
            - 417
            - 418
            - 421
            - 422
            - 423
            - 424
            - 428
            - 429
            - 456
            - 500
            - 501
            - 502
            - 503
            - 504
            - 505
            - 507
            - 508
        errorCode:
          description: The application specific error code
          type: string
        message:
          description: The error message
          type: string
        errors:
          description: Zod validation errors
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
            required:
              - field
              - message
              - code
      required:
        - statusCode
        - errorCode
        - message
...
    CarResponseDto:
      type: object
      properties:
        apiVersion:
          type: string
          description: The version of the API
          example: 1.0.0
        data:
          description: The data of the response
          allOf:
            - $ref: "#/components/schemas/CarDto"
      required:
        - apiVersion
        - data
...
    CarModelResponseDto:
      type: object
      properties:
        apiVersion:
          type: string
          description: The version of the API
          example: 1.0.0
        data:
          description: The data of the response
          allOf:
            - $ref: "#/components/schemas/CarModelDto"
      required:
        - apiVersion
        - data
...